+++ tests/testthat/test-colwise-select.R
@@ -167,3 +167,11 @@ test_that("mutate_all does not change the order of columns (#3351)", {
   tbl <- group_by(tibble(x = 1:4, y = 1:4, z = 1:4), y)
   expect_message(expect_identical(names(mutate_all(tbl, identity)), names(tbl)), "ignored")
 })
+
+test_that("select_at(.strict = FALSE) is forgiving (#4111)", {
+  res <- expect_warning(select_at(iris, c("Sepal.Length", "foo", "bar", "Species"), .strict = FALSE), NA)
+  expect_identical(
+    res,
+    select(iris, Sepal.Length, Species)
+  )
+})